fix: move spinner to styles and fix todo
