+++ src/styles/index.ts
@@ -2,3 +2,4 @@ export { default as H2 } from './H2';
 export { default as Section } from './Section';
 export { default as P } from './P';
 export { default as Skeleton } from './Skeleton';
+export { default as Spinner } from './Spinner';
